+++ test/test_cli.py
@@ -42,20 +42,20 @@ def execute_cmd(cmd: Union[str, List[str]], raise_error: bool = True) -> Tuple[i
     debug_stdout = collections.deque(maxlen=1000)
     print("\n" + "*" * 10 + "  execute_cmd starts  " + "*" * 10 + "\n")
     while 1:
-        line = s.stdout.readline().decode("utf-8", errors="replace")
+        line = s.stdout.readline().decode("utf-8", errors="replace")  # type: ignore
         if line.rstrip():
             debug_stdout.append(line)
             # print(line)
         exitcode = s.poll()
         if exitcode is not None:
-            for line in s.stdout:
+            for line in s.stdout:  # type: ignore
                 debug_stdout.append(line.decode("utf-8", errors="replace"))
             if exitcode is not None and exitcode != 0:
                 error_msg = " ".join(cmd) if not isinstance(cmd, str) else cmd
                 error_msg += "\n"
                 error_msg += "".join(debug_stdout)
                 s.communicate()
-                s.stdout.close()
+                s.stdout.close()  # type: ignore
                 print("\n" + "*" * 10 + "  execute_cmd failed  " + "*" * 10 + "\n")
                 if raise_error:
                     raise subprocess.CalledProcessError(exitcode, error_msg)
@@ -65,7 +65,7 @@ def execute_cmd(cmd: Union[str, List[str]], raise_error: bool = True) -> Tuple[i
             else:
                 break
     s.communicate()
-    s.stdout.close()
+    s.stdout.close()  # type: ignore
     output_msg = list(debug_stdout)
 
     print("\n" + "*" * 10 + "  execute_cmd succeeded  " + "*" * 10 + "\n")